Factor login page and form, and allow direct authentication when it makes sense.
auth* plugins can now provide a redirection link if their authentication
requires no interaction on the forge (eg. with CAS, Kerberos or other
single-sign-on systems).  When there's only one active auth* plugin, and it
provides such a link, the "log in" link can be made to point at wherever the
user needs to be redirected, which avoids a form with only one button and
no parameters.


git-svn-id: svn+ssh://scm.fusionforge.org/srv/svn/fusionforge/trunk@12809 9d84d37e-dcb1-4aad-b103-6f3d92f53bf6